Update server
- add state
- update test to close server
- refactor server directory
- import server index into bot index
