--- server/routes/invites.js
@@ -1,7 +1,0 @@
-const router = require("express").Router();
-
-router.get("/:id", (req, res) => {
-  res.redirect(process.env.URL);
-});
-
-module.exports = router;
